Shorten README to less noisy cribnotes-style summaries
Also moving the notes on why certain names are included to
comments in the scripts themselves.

The idea is, at this point, you expect envigor to "just work",
and if you're checking the readme, it's to check that it provides
the interface you're looking for. If you need to determine the
exact variables it's going to pull from (which we're now moving
toward being more all-encompassing with our search for) and how it
assembles these values, your query is complex and specific enough
that you should just read the logic from the code.